fix: track file renaming
